commit 7

+++ Drivers/TIMER.h
@@ -17,7 +17,7 @@ void MyTimer_ActiveIT ( TIM_TypeDef * Timer, char Prio, void ( * IT_function ) (
 // Activation d'une interruption
 // La fonction MyTimer_Base_Init doit avoir été lancée avant de configurer une interruption
 // Paramètres : - TIM_TypeDef * Timer : Timer concerné
-//							- char Prio : Priorité allant de 0 à 15 (0 étant la plus grande priorité)
+//              - char Prio : Priorité allant de 0 à 15 (0 étant la plus grande priorité)
 
 #define MyTimer_Base_Start(Timer) Timer->CR1 |= TIM_CR1_CEN
 #define MyTimer_Base_Stop(Timer) Timer->CR1 &= ~(TIM_CR1_CEN)
@@ -28,7 +28,7 @@ void MyTimer_PWM_Init_Channel(TIM_TypeDef * Timer, char Channel);
 // Paramètres : - TIM_TypeDef * Timer : Timer concerné
 //							- char Channel : canal du Timer (1, 2, 3 ou 4)
 
-void MyTimer_PWM_Set_Rapport_Cyclique(TIM_TypeDef * Timer, char Channel, int Rapport_Cyclique);
+void MyTimer_PWM_Set_Rapport_Cyclique(TIM_TypeDef * Timer, char Channel, double Rapport_Cyclique);
 // Affectation d'un rapport cyclique à une PWM d'un canal particulier d'un Timer donné
 
 #endif
